requirements.txt
 pystray~=0.19.5
pillow~=10.3.0
plyer~=2.1.00123456789ABCDEFGHIJKLMNOPQRSTUVWXY0123456789*+-./